+++ src/stories/screens/Settings/index.js
@@ -113,6 +113,8 @@ class Settings extends React.Component {
       url,
       user_name,
       password,
+      isHttps,
+      toggleHttps,
 
       // Attendant
       attendant,
@@ -211,6 +213,8 @@ class Settings extends React.Component {
           url={url}
           user_name={user_name}
           password={password}
+          isHttps={isHttps}
+          toggleIsHttps={toggleHttps}
         />
       );
     }